implement disable of 'Models' section
The section of 'Models' always is created if there is any model
used during the Api definition, with this change we enable the
tool to not to add this section if the user wants.
'models' parameter which default value is true indicates whether
you want the models section to be generated or not, if you pass the
'-models=false' parameter then the models section will not bee added
to the generated output.

Signed-off-by: Leoswaldo Macias <lmaciasm10@gmail.com>