vendoring: make vendoring to avoid looking into hidden dirs and vendor itself
There is a "standard" that allows project to add non pertinent
code of the project itself by creating directories with naming convention
starting with "_". This directories were being listed during the
inspection of import modules, while it should not.

This commit has the fix for the previous escenario and also to avoid
looking into vendor directory itself.

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>