vendoring: add flag to specify the vendor directory
As today the if we enable vendoring through the GO15VENDOREXPERIMENT
variable, swagger will look for the vendor directory from the current
working directory, so in case your vendored dependencies are in a
different directory than where you are executing it it will not find
the, to prevent it, this commit adds the flag vendoringPath so the
users can specify the directory of their vendoring even if its outside
their project. The tool will keep its default behavior if the
vendoringPath flag is not provided.

Signed-off-by: Leoswaldo Macias <leoswaldo.macias@intel.com>